refactor: move owner populating to ready event
